Define typedef aliases for selst and tokenst.

+++ scr_delete_characters.h
@@ -4,7 +4,9 @@
 #ifndef SCR_DEL_CHARS_H
 #define SCR_DEL_CHARS_H
 
+#include <stdint.h>
+
 //  Delete count characters from the current position.
-void scr_delete_characters(int count);
+void scr_delete_characters(uint8_t count);
 
 #endif//SCR_DEL_CHARS_H
